cast ra, dec, zqso to float64 when reading catalog

+++ py/picca/io.py
@@ -48,15 +48,15 @@ def read_drq(drq,zmin,zmax,keep_bal,bi_max=None):
 
     ## Redshift
     try:
-        zqso = vac[1]["Z"][:]
+        zqso = vac[1]["Z"][:].astype('float64')
     except:
         sys.stderr.write("Z not found (new DRQ >= DRQ14 style), using Z_VI (DRQ <= DRQ12)\n")
-        zqso = vac[1]["Z_VI"][:]
+        zqso = vac[1]["Z_VI"][:].astype('float64')
 
     ## Info of the primary observation
     thid  = vac[1]["THING_ID"][:]
-    ra    = vac[1]["RA"][:]
-    dec   = vac[1]["DEC"][:]
+    ra    = vac[1]["RA"][:].astype('float64')
+    dec   = vac[1]["DEC"][:].astype('float64')
     plate = vac[1]["PLATE"][:]
     mjd   = vac[1]["MJD"][:]
     fid   = vac[1]["FIBERID"][:]
